cdk/requirements.txt
aws-cdk-lib==2.130.0
constructs>=10.0.0,<11.0.0
# aws-cdk.aws-iam
# aws-cdk.aws-eks
# aws-cdk.aws-ec2
# aws-cdk.aws-ssm
aws_cdk.lambda_layer_kubectl_v29
# aws_cdk.aws_ssm
# aws-cdk.core

